Refining the src struture: shorter component integration
