--- src/components/Client/Client.js
@@ -1,8 +1,0 @@
-import React from "react"
-export default function About() {
-  return (
-  <div>
-      <h1>Client Page</h1>
-  </div>
-  )
-}
