+++ src/components/Client/index.js
@@ -1,0 +1,9 @@
+import React from "react"
+export default function Client() {
+  return (
+  <div>
+      <p style={{padding:"10px", margin:"20px", backgroundColor:"yellow", textAlign:"center"}}>
+         I am <strong>CLIENT</strong> page's body.</p>
+  </div>
+  )
+}
