+++ src/components/Contact/index.js
@@ -1,0 +1,9 @@
+import React from "react"
+export default function Contact() {
+  return (
+  <div>
+      <p style={{padding:"10px", margin:"20px", backgroundColor:"yellow", textAlign:"center"}}>
+         I am <strong>CONTACT</strong> page's body.</p>
+  </div>
+  )
+}
