--- src/components/Home/index.js
@@ -1,4 +1,0 @@
-export { default as HowAreWeDifferentSection } from './HowAreWeDifferentSection';
-export { default as TheBottomLineSection } from './TheBottomLineSection';
-export { default as OurRootsSection } from './OurRootsSection';
-export {default as TimelineSection} from './TimelineSection';
