--- src/components/Partners/Partners.js
@@ -1,8 +1,0 @@
-import React from "react"
-export default function Partners() {
-  return (
-  <div>
-       <h1>Partners page</h1>
-    </div>
-  )
-}
