--- src/components/Sponsors/Sponsors.js
@@ -1,8 +1,0 @@
-import React from "react"
-export default function Sponsors() {
-  return (
-  <div>
-       <h1>Sponsors page</h1>
-    </div>
-  )
-}
