+++ src/components/Sponsors/index.js
@@ -1,0 +1,9 @@
+import React from "react"
+export default function SponsorsPage() {
+  return (
+  <div>
+      <p style={{padding:"10px", margin:"20px", backgroundColor:"yellow", textAlign:"center"}}>
+         I am <strong>SPONSORS</strong> page's body.</p>
+  </div>
+  )
+}
