--- src/components/Volunteer/Volunteer.js
@@ -1,8 +1,0 @@
-import React from "react"
-export default function Volunteer() {
-  return (
-  <div>
-      <h1>Volunteer Page</h1>
-  </div>
-  )
-}
