+++ src/pages/client/index.js
@@ -1,10 +1,11 @@
 import React from "react"
 import { MainLayout } from "../../layout";
+import PageBody from "../../components/Client/"
 
 export default function ClientPage() {
   return (
     <MainLayout>
-      <p>This is Client Page</p>
+      <PageBody/>
     </MainLayout>
   )
 }
