+++ src/pages/sponsors/index.js
@@ -1,10 +1,11 @@
 import React from "react"
 import { MainLayout } from "../../layout";
+import PageBody from "../../components/Sponsors"
 
-export default function SponsorPage() {
+export default function SponsorsPage() {
   return (
     <MainLayout>
-      <p>This is Sponsor Page</p>
+      <PageBody/>
     </MainLayout>
   )
 }
