+++ src/pages/volunteer/index.js
@@ -1,10 +1,11 @@
 import React from "react"
 import { MainLayout } from "../../layout";
+import PageBody from "../../components/Volunteer"
 
 export default function VolunteerPage() {
   return (
     <MainLayout>
-      <p>This is Volunteer Page</p>
+      <PageBody/>
     </MainLayout>
   )
 }
